requirements.txt
streamlit
langchain
langchain-community
langchain-groq
langchain-ollama
faiss-cpu
sentence-transformers
transformers
torch
firecrawl-py
python-dotenv
matplotlib
requests
beautifulsoup4
wikipedia
arxiv